+++ steel/src/config.rs
@@ -17,41 +17,8 @@ use std::collections::BTreeMap;
 
 use alloy_primitives::{BlockNumber, ChainId};
 use anyhow::bail;
-use once_cell::sync::Lazy;
 use revm::primitives::SpecId;
 use serde::{Deserialize, Serialize};
-
-/// The Ethereum Mainnet specification.
-pub static ETH_MAINNET_CHAIN_SPEC: Lazy<ChainSpec> = Lazy::new(|| ChainSpec {
-    chain_id: 1,
-    max_spec_id: SpecId::CANCUN,
-    hard_forks: BTreeMap::from([
-        (SpecId::MERGE, ForkCondition::Block(15537394)),
-        (SpecId::SHANGHAI, ForkCondition::Timestamp(1681338455)),
-        (SpecId::CANCUN, ForkCondition::Timestamp(1710338135)),
-    ]),
-    gas_constants: BTreeMap::from([(SpecId::LONDON, EIP1559_CONSTANTS_DEFAULT)]),
-});
-
-/// The Ethereum Sepolia specification.
-pub static ETH_SEPOLIA_CHAIN_SPEC: Lazy<ChainSpec> = Lazy::new(|| ChainSpec {
-    chain_id: 11155111,
-    max_spec_id: SpecId::CANCUN,
-    hard_forks: BTreeMap::from([
-        (SpecId::MERGE, ForkCondition::Block(1735371)),
-        (SpecId::SHANGHAI, ForkCondition::Timestamp(1677557088)),
-        (SpecId::CANCUN, ForkCondition::Timestamp(1706655072)),
-    ]),
-    gas_constants: BTreeMap::from([(SpecId::LONDON, EIP1559_CONSTANTS_DEFAULT)]),
-});
-
-/// The gas constants as defined in [EIP-1559](https://eips.ethereum.org/EIPS/eip-1559).
-pub const EIP1559_CONSTANTS_DEFAULT: Eip1559Constants = Eip1559Constants {
-    base_fee_change_denominator: 8,
-    base_fee_max_increase_denominator: 8,
-    base_fee_max_decrease_denominator: 8,
-    elasticity_multiplier: 2,
-};
 
 /// The condition at which a fork is activated.
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
@@ -67,6 +34,7 @@ pub enum ForkCondition {
 
 impl ForkCondition {
     /// Returns whether the condition has been met.
+    #[inline]
     pub fn active(&self, block_number: BlockNumber, timestamp: u64) -> bool {
         match self {
             ForkCondition::Block(block) => *block <= block_number,
@@ -76,36 +44,24 @@ impl ForkCondition {
     }
 }
 
-/// [EIP-1559](https://eips.ethereum.org/EIPS/eip-1559) parameters.
-#[derive(Debug, Clone, Copy, Eq, PartialEq, Serialize, Deserialize)]
-pub struct Eip1559Constants {
-    pub base_fee_change_denominator: u64,
-    pub base_fee_max_increase_denominator: u64,
-    pub base_fee_max_decrease_denominator: u64,
-    pub elasticity_multiplier: u64,
-}
-
 /// Specification of a specific chain.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ChainSpec {
-    chain_id: ChainId,
-    max_spec_id: SpecId,
-    hard_forks: BTreeMap<SpecId, ForkCondition>,
-    gas_constants: BTreeMap<SpecId, Eip1559Constants>,
+    /// Chain identifier.
+    pub chain_id: ChainId,
+    /// ID of the latest supported revm specification.
+    pub max_spec_id: SpecId,
+    /// Map revm specification IDs to their respective activation condition.
+    pub hard_forks: BTreeMap<SpecId, ForkCondition>,
 }
 
 impl ChainSpec {
     /// Creates a new configuration consisting of only one specification ID.
-    pub fn new_single(
-        chain_id: ChainId,
-        spec_id: SpecId,
-        eip_1559_constants: Eip1559Constants,
-    ) -> Self {
+    pub fn new_single(chain_id: ChainId, spec_id: SpecId) -> Self {
         ChainSpec {
             chain_id,
             max_spec_id: spec_id,
             hard_forks: BTreeMap::from([(spec_id, ForkCondition::Block(0))]),
-            gas_constants: BTreeMap::from([(spec_id, eip_1559_constants)]),
         }
     }
     /// Returns the network chain ID.
@@ -137,13 +93,6 @@ impl ChainSpec {
             None => bail!("no supported fork for block {}", block_number),
         }
     }
-    /// Returns the Eip1559 constants for a given [SpecId].
-    pub fn gas_constants(&self, spec_id: SpecId) -> Option<&Eip1559Constants> {
-        self.gas_constants
-            .range(..=spec_id)
-            .next_back()
-            .map(|(_, v)| v)
-    }
 
     fn spec_id(&self, block_number: BlockNumber, timestamp: u64) -> Option<SpecId> {
         for (spec_id, fork) in self.hard_forks.iter().rev() {
@@ -159,6 +108,18 @@ impl ChainSpec {
 mod tests {
     use super::*;
 
+    use once_cell::sync::Lazy;
+
+    static ETH_MAINNET_CHAIN_SPEC: Lazy<ChainSpec> = Lazy::new(|| ChainSpec {
+        chain_id: 1,
+        max_spec_id: SpecId::CANCUN,
+        hard_forks: BTreeMap::from([
+            (SpecId::MERGE, ForkCondition::Block(15537394)),
+            (SpecId::SHANGHAI, ForkCondition::Timestamp(1681338455)),
+            (SpecId::CANCUN, ForkCondition::Timestamp(1710338135)),
+        ]),
+    });
+
     #[test]
     fn spec_id() {
         assert_eq!(ETH_MAINNET_CHAIN_SPEC.spec_id(15537393, 0), None);
@@ -173,19 +134,6 @@ mod tests {
         assert_eq!(
             ETH_MAINNET_CHAIN_SPEC.spec_id(0, 1681338455),
             Some(SpecId::SHANGHAI)
-        );
-    }
-
-    #[test]
-    fn gas_constants() {
-        assert_eq!(ETH_MAINNET_CHAIN_SPEC.gas_constants(SpecId::BERLIN), None);
-        assert_eq!(
-            ETH_MAINNET_CHAIN_SPEC.gas_constants(SpecId::MERGE),
-            Some(&EIP1559_CONSTANTS_DEFAULT)
-        );
-        assert_eq!(
-            ETH_MAINNET_CHAIN_SPEC.gas_constants(SpecId::SHANGHAI),
-            Some(&EIP1559_CONSTANTS_DEFAULT)
         );
     }
 }
