+++ steel/src/lib.rs
@@ -15,13 +15,10 @@
 #![cfg_attr(not(doctest), doc = include_str!("../README.md"))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
-use std::{fmt::Debug, rc::Rc};
+use std::fmt::Debug;
 
 use ::serde::{Deserialize, Serialize};
-use alloy_primitives::{
-    b256, keccak256, Address, BlockNumber, Bytes, Sealable, Sealed, TxNumber, B256, U256,
-};
-use alloy_rlp_derive::{RlpDecodable, RlpEncodable};
+use alloy_primitives::{BlockNumber, Bytes, Sealable, Sealed, B256, U256};
 use revm::primitives::{BlockEnv, CfgEnvWithHandlerCfg, HashMap, SpecId};
 
 pub mod config;
@@ -31,6 +28,7 @@ pub mod ethereum;
 pub mod host;
 mod mpt;
 pub mod serde;
+mod state;
 
 pub use contract::{CallBuilder, Contract};
 pub use mpt::MerkleTrie;
@@ -100,6 +98,7 @@ mod private {
 
 /// Solidity struct representing the committed block used for validation.
 pub use private::Commitment as SolCommitment;
+use state::StateDb;
 
 /// Alias for readability, do not make public.
 pub(crate) type GuestEvmEnv<H> = EvmEnv<StateDb, H>;
@@ -142,99 +141,9 @@ impl<D, H: EvmBlockHeader> EvmEnv<D, H> {
     }
 
     /// Returns the header of the environment.
+    #[inline]
     pub fn header(&self) -> &H {
         self.header.inner()
-    }
-}
-
-/// A simple read-only EVM database.
-///
-/// It is backed by a single [MerkleTrie] for the accounts and one [MerkleTrie] each for the
-/// accounts' storages. It panics when data is queried that is not contained in the tries.
-pub struct StateDb {
-    state_trie: MerkleTrie,
-    storage_tries: HashMap<B256, Rc<MerkleTrie>>,
-    contracts: HashMap<B256, Bytes>,
-    block_hashes: HashMap<u64, B256>,
-}
-
-impl StateDb {
-    /// Creates a new state database from the given tries.
-    pub fn new(
-        state_trie: MerkleTrie,
-        storage_tries: impl IntoIterator<Item = MerkleTrie>,
-        contracts: impl IntoIterator<Item = Bytes>,
-        block_hashes: HashMap<u64, B256>,
-    ) -> Self {
-        let contracts = contracts
-            .into_iter()
-            .map(|code| (keccak256(&code), code))
-            .collect();
-        let storage_tries = storage_tries
-            .into_iter()
-            .map(|trie| (trie.hash_slow(), Rc::new(trie)))
-            .collect();
-        Self {
-            state_trie,
-            contracts,
-            storage_tries,
-            block_hashes,
-        }
-    }
-
-    fn account(&self, address: Address) -> Option<StateAccount> {
-        self.state_trie
-            .get_rlp(keccak256(address))
-            .expect("invalid state value")
-    }
-
-    fn code_by_hash(&self, hash: B256) -> &Bytes {
-        self.contracts
-            .get(&hash)
-            .unwrap_or_else(|| panic!("code not found: {}", hash))
-    }
-
-    fn block_hash(&self, number: U256) -> B256 {
-        // block number is never bigger then u64::MAX
-        let number: u64 = number.to();
-        let hash = self
-            .block_hashes
-            .get(&number)
-            .unwrap_or_else(|| panic!("block not found: {}", number));
-        *hash
-    }
-
-    fn storage_trie(&self, root: &B256) -> Option<&Rc<MerkleTrie>> {
-        self.storage_tries.get(root)
-    }
-}
-
-/// Hash of an empty byte array, i.e. `keccak256([])`.
-pub const KECCAK_EMPTY: B256 =
-    b256!("c5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470");
-
-/// Represents an account within the state trie.
-#[derive(Debug, Clone, PartialEq, Eq, RlpEncodable, RlpDecodable)]
-struct StateAccount {
-    /// The number of transactions sent from this account's address.
-    pub nonce: TxNumber,
-    /// The number of Wei owned by this account's address.
-    pub balance: U256,
-    /// The root of the account's storage trie.
-    pub storage_root: B256,
-    /// The hash of the EVM code of this account.
-    pub code_hash: B256,
-}
-
-impl Default for StateAccount {
-    /// Provides default values for a [StateAccount].
-    fn default() -> Self {
-        Self {
-            nonce: 0,
-            balance: U256::ZERO,
-            storage_root: mpt::EMPTY_ROOT_HASH,
-            code_hash: KECCAK_EMPTY,
-        }
     }
 }
